+++ core/server/utils/index.js
@@ -104,7 +104,6 @@ utils = {
     },
 
     readCSV: require('./read-csv'),
-    readDirectory: require('./read-directory'),
     removeOpenRedirectFromUrl: require('./remove-open-redirect-from-url'),
     zipFolder: require('./zip-folder'),
     generateAssetHash: require('./asset-hash'),
