+++ core/test/unit/utils/packages_spec.js
@@ -1,0 +1,250 @@
+var should = require('should'),
+    tmp = require('tmp'),
+    join = require('path').join,
+    fs = require('fs'),
+
+    // Things we are testing
+    packages = require('../../../server/utils/packages'),
+    parsePackageJson = packages.parsePackageJSON,
+    readPackages = packages.readPackages;
+
+describe('Package Utils', function () {
+    describe('Parse Package.json', function () {
+        it('should parse valid package.json', function (done) {
+            var pkgJson, tmpFile;
+
+            tmpFile = tmp.fileSync();
+            pkgJson = JSON.stringify({
+                name: 'test',
+                version: '0.0.0'
+            });
+
+            fs.writeSync(tmpFile.fd, pkgJson);
+
+            parsePackageJson(tmpFile.name)
+                .then(function (pkg) {
+                    pkg.should.eql({
+                        name: 'test',
+                        version: '0.0.0'
+                    });
+
+                    done();
+                })
+                .catch(done)
+                .finally(tmpFile.removeCallback);
+        });
+
+        it('should fail when name is missing', function (done) {
+            var pkgJson, tmpFile;
+
+            tmpFile = tmp.fileSync();
+            pkgJson = JSON.stringify({
+                version: '0.0.0'
+            });
+
+            fs.writeSync(tmpFile.fd, pkgJson);
+
+            parsePackageJson(tmpFile.name)
+                .then(function () {
+                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
+                })
+                .catch(function (err) {
+                    err.message.should.equal('"name" or "version" is missing from theme package.json file.');
+                    err.context.should.equal(tmpFile.name);
+                    err.help.should.equal('This will be required in future. Please see http://docs.ghost.org/themes/');
+
+                    done();
+                })
+                .catch(done)
+                .finally(tmpFile.removeCallback);
+        });
+
+        it('should fail when version is missing', function (done) {
+            var pkgJson, tmpFile;
+
+            tmpFile = tmp.fileSync();
+            pkgJson = JSON.stringify({
+                name: 'test'
+            });
+
+            fs.writeSync(tmpFile.fd, pkgJson);
+
+            parsePackageJson(tmpFile.name)
+                .then(function () {
+                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
+                })
+                .catch(function (err) {
+                    err.message.should.equal('"name" or "version" is missing from theme package.json file.');
+                    err.context.should.equal(tmpFile.name);
+                    err.help.should.equal('This will be required in future. Please see http://docs.ghost.org/themes/');
+
+                    done();
+                })
+                .catch(done)
+                .finally(tmpFile.removeCallback);
+        });
+
+        it('should fail when JSON is invalid', function (done) {
+            var pkgJson, tmpFile;
+
+            tmpFile = tmp.fileSync();
+            pkgJson = '{name:"test"}';
+
+            fs.writeSync(tmpFile.fd, pkgJson);
+
+            parsePackageJson(tmpFile.name)
+                .then(function () {
+                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
+                })
+                .catch(function (err) {
+                    err.message.should.equal('Theme package.json file is malformed');
+                    err.context.should.equal(tmpFile.name);
+                    err.help.should.equal('This will be required in future. Please see http://docs.ghost.org/themes/');
+
+                    done();
+                })
+                .catch(done)
+                .finally(tmpFile.removeCallback);
+        });
+
+        it('should fail when file is missing', function (done) {
+            var tmpFile = tmp.fileSync();
+
+            tmpFile.removeCallback();
+            parsePackageJson(tmpFile.name)
+                .then(function () {
+                    done(new Error('parsePackageJson succeeded, but should\'ve failed'));
+                })
+                .catch(function (err) {
+                    err.message.should.equal('Could not read package.json file');
+                    err.context.should.equal(tmpFile.name);
+
+                    done();
+                })
+                .catch(done);
+        });
+    });
+
+    describe('Read Packages', function () {
+        it('should read directory recursively', function (done) {
+            var themePath = tmp.dirSync({unsafeCleanup: true});
+
+            // create example theme
+            fs.mkdirSync(join(themePath.name, 'partials'));
+            fs.writeFileSync(join(themePath.name, 'index.hbs'));
+            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
+
+            readPackages(themePath.name)
+                .then(function (tree) {
+                    tree.should.eql({
+                        partials: {
+                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
+                        },
+                        'index.hbs': join(themePath.name, 'index.hbs')
+                    });
+
+                    done();
+                })
+                .catch(done)
+                .finally(themePath.removeCallback);
+        });
+
+        it('should read directory and ignore unneeded items', function (done) {
+            var themePath = tmp.dirSync({unsafeCleanup: true});
+
+            // create example theme
+            fs.mkdirSync(join(themePath.name, 'partials'));
+            fs.writeFileSync(join(themePath.name, 'index.hbs'));
+            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
+
+            // create some trash
+            fs.mkdirSync(join(themePath.name, 'node_modules'));
+            fs.mkdirSync(join(themePath.name, 'bower_components'));
+            fs.mkdirSync(join(themePath.name, '.git'));
+            fs.writeFileSync(join(themePath.name, '.DS_Store'));
+
+            readPackages(themePath.name, {ignore: ['.git']})
+                .then(function (tree) {
+                    tree.should.eql({
+                        partials: {
+                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
+                        },
+                        'index.hbs': join(themePath.name, 'index.hbs')
+                    });
+
+                    done();
+                })
+                .catch(done)
+                .finally(themePath.removeCallback);
+        });
+
+        it('should read directory and parse package.json files', function (done) {
+            var themePath, pkgJson;
+
+            themePath = tmp.dirSync({unsafeCleanup: true});
+            pkgJson = JSON.stringify({
+                name: 'test',
+                version: '0.0.0'
+            });
+
+            // create example theme
+            fs.mkdirSync(join(themePath.name, 'partials'));
+            fs.writeFileSync(join(themePath.name, 'package.json'), pkgJson);
+            fs.writeFileSync(join(themePath.name, 'index.hbs'));
+            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
+
+            readPackages(themePath.name)
+                .then(function (tree) {
+                    tree.should.eql({
+                        partials: {
+                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
+                        },
+                        'index.hbs': join(themePath.name, 'index.hbs'),
+                        'package.json': {
+                            name: 'test',
+                            version: '0.0.0'
+                        }
+                    });
+
+                    done();
+                })
+                .catch(done)
+                .finally(themePath.removeCallback);
+        });
+
+        it('should read directory and ignore invalid package.json files', function (done) {
+            var themePath, pkgJson;
+
+            themePath = tmp.dirSync({unsafeCleanup: true});
+            pkgJson = JSON.stringify({
+                name: 'test'
+            });
+
+            // create example theme
+            fs.mkdirSync(join(themePath.name, 'partials'));
+            fs.writeFileSync(join(themePath.name, 'package.json'), pkgJson);
+            fs.writeFileSync(join(themePath.name, 'index.hbs'));
+            fs.writeFileSync(join(themePath.name, 'partials', 'navigation.hbs'));
+
+            readPackages(themePath.name)
+                .then(function (tree) {
+                    tree.should.eql({
+                        partials: {
+                            'navigation.hbs': join(themePath.name, 'partials', 'navigation.hbs')
+                        },
+                        'index.hbs': join(themePath.name, 'index.hbs'),
+                        'package.json': null
+                    });
+
+                    done();
+                })
+                .catch(done)
+                .finally(themePath.removeCallback);
+        });
+    });
+
+    describe('Filter Packages', function () {
+        it('should filter packages correctly?');
+    });
+});
+
